gitaly: Use gRPC liveness probe
Gitaly exposes an unauthenticated gRPC healthcheck service [1]. The
healthcheck becomes healthy after the gRPC server is ready to accept
requests, which is more reliable than the current healthcheck script
that checks for an open port.

Update the various Gitaly probes to use gRPC liveness probes [2]
targeting the Service port. This should be an internal, no-op change as
far as users are concerned, but has a few benefits for us:

1. Calling the gRPC healthcheck service provides a more accurate
   indication of health.
2. The healthcheck service can be extended independently of how Gitaly
   is deployed; so additional checks can be added transparently in the
   future.

[1] https://gitlab.com/gitlab-org/gitaly/-/issues/5776
[2] https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/#define-a-grpc-liveness-probe