Introduce minimal AutoFlow support with Temporal Cloud
This change set introduces minimal support for AutoFlow specifically
with Temporal Cloud. It does not support anything else at this time.

This change set is required for the GitLab internal-use only deployment
of the first AutoFlow experiment deployment.
See https://gitlab.com/groups/gitlab-org/-/epics/16181 for context.
It's the most boring solution to deploy AutoFlow.

This is not meant to be enabled by our customers and is disabled by
default in the chart.

Nonetheless, this change set includes minimal documentation in order for
GitLab to know how to configure AutoFlow.

Refs https://gitlab.com/gitlab-org/gitlab/-/issues/508259