kas: fix nginx conditionals of Ingress
Adjust the conditional in the Ingress template so that
all annotations for NGINX are guarded. This improves the
previous state, where one level of depth was missed,
resulting in the presence of some annotations that are
not needed outside NGINX as a Ingress provider. Their
presence was innocuous, but confusing to users of the
chart.

changelog: fixed