Remove flag to toggle Redis in KAS chart
The gitlab.kas.redis.enabled value was a misleading key,
as Redis has become a hard dependency for the GitLab kas
subchart. I.e. it's not possible to use this subchart with
Redis disabled. So we're removing this key from our GitLab
kas subchart values.

If you have it set to true, feel free to remove it from your
custom helm values file as well, as it's a no-op.

Changelog: other