gitaly: Add bundleUri configuration block
The Bundle URIs functionality [1] was recently introduced into Gitaly
and is ready for testing. Since some repositories in `gstg` are hosted
by Gitaly pods running on Kubernetes, the new configuration keys must
be explicitly added to the ConfigMap in the Gitaly Chart. On top of
testing, this new key will be required for customer installations too.

Add a bundleUri section to expose these configuration keys.

[1] https://docs.gitlab.com/administration/gitaly/bundle_uris/