Go services: Drop literal address
Workhorse, Pages, and Gitaly set a listen address of `0.0.0.0`,
for the TCP network, which still results in listening on IPv4 and
IPv6 interfaces:

>>>
For TCP networks, if the host in the address parameter is empty or a
literal unspecified IP address, Listen listens on all available unicast
and anycast IP addresses of the local system.
>>>

https://pkg.go.dev/net#Listen

We can drop this and leave the address empty, to avoid any confusion
about the IPv6 behaviour.