Make it possible to wait for migrations container to complete
We would like to run some `extraContainers` which can start running only after the main migrations
container has completed running successfully. This change adds a volume (following the `toolbox-tmp`
pattern) and updates the command of the migrations container to write a "completion" file to the
`toolbox-tmp` volume. Extra containers can wait for the existence of this file, if they are
interested in the status of the migrations job.

Changelog: changed