Make runner registration token an optional secret
With this change the migration jobs won't anymore
require an existing runner token secret to start.

This is inline with expectations that this registration token
created by the shared secrets jobs should not be used anymore
as it's deprecated. We don't expect this to affect anyone's
workflow. But if you're deploying a chart with the shared secrets
job disabled, or without the runner embedded chart, you could
previously be getting pod failures on your migration jobs. From
now on, the migration job is expected to complete successfully
also for these cases.

Changelog: changed