Include HTTPS in the configuration of GitLab-Migrations job
GitLab's codebase infers the fully qualified URL that the instance is running at using three
components of the configuration within the `config/gitlab.yml` file: `host`, `https`, and `port`, in
the [`build_gitlab_url`] function.

The logic for determining the URL in rough pseudocode is:

```
{{ https ? "https" : "http" }}://{{ host }}{{ is_standard_port_for_protocol?(port) ? "" : (":" + port) }}
```

The existing GitLab configuration used by the Migrations chart does not contain the `https`
key. This implies that the URL is inferred as `http://`, which is incorrect.

It is causing a problem for `gstg` (`staging.gitlab.com`) where the [`Gitlab.com?` check] within the
GitLab rails codebase is reporting `false`, even though we expect it to be `true`. I expect the same
problem to happen in `gprd` as well, because there we are running on `https://gitlab.com`.

This patch includes the `https` key in the `gitlab.yml` configuration file. The changed line is
copied from the [`toolbox/templates/configmap.yml`] file, which has a far more complete
`gitlab.yml.erb` file.

[`build_gitlab_url`]: https://gitlab.com/gitlab-org/gitlab/blob/14ffa9418021d544d9a6fe57c9c3ae104bd5e94c/config/settings.rb#L63-66
[`Gitlab.com?` check]: https://gitlab.com/gitlab-org/gitlab/blob/14ffa9418021d544d9a6fe57c9c3ae104bd5e94c/lib/gitlab.rb#L96-99
[`toolbox/templates/configmap.yml`]: https://gitlab.com/gitlab-org/charts/gitlab/-/blob/3df625579821dd42b7a4b283beab131ff5158718/charts/gitlab/charts/toolbox/templates/configmap.yaml