Add config to disable other components and only execute migrations
This adds a `registry.api.enabled` value that can be used to disable
the deployment, hpa, pdb, and service resources in the registry chart.

This will be useful for executing Container Registry post deployment
migrations after an application deployment, as a separate Helm
release, since it allows us to disable resources that are not required
for executing post deployment migrations.