feat: Run ClickHouse migrations in GitLab-Migrations chart
Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/21436.

Running ClickHouse migrations using our public Helm charts is not supported at the moment. The
[documentation for ClickHouse migrations] recommends running `gitlab-rake` directly. We don't want
to run `gitlab-rake` directly on Toolbox pods for GitLab.com. The GitLab-Migrations chart already
runs the [/scripts/db-migrate][1] script. [/scripts/db-migrate][1] runs the [Rake task
gitlab:db:configure][2]. This Rake task invokes the [gitlab:clickhouse:migrate Rake task][3] in
turn.

So, all we need to do is update the ConfigMap and mount the right secrets in the `_jobspec` of the
migrations Job. ClickHouse is configured in the ConfigMap for Webservice, Sidekiq, and Toolbox. I
have copied over that snippet of configuration to the migrations Job.

[documentation for ClickHouse migrations]: https://docs.gitlab.com/integration/clickhouse/#run-clickhouse-migrations

[1]: https://gitlab.com/gitlab-org/build/CNG/-/blob/6f4869fe6e56ed1563232573706e01618df792c5/gitlab-rails/scripts/db-migrate#L22
[2]: https://gitlab.com/gitlab-org/gitlab/blob/316892ec48dfae3a81f8b3f85eecd76078cf3edc/lib/tasks/gitlab/db.rake#L100-105
[3]: https://gitlab.com/gitlab-org/gitlab/blob/316892ec48dfae3a81f8b3f85eecd76078cf3edc/lib/tasks/gitlab/db.rake#L175-178

Changelog: added