Disable Workhorse preStop hook if health checker listener is enabled
We do not want to use a `preStop` hook with Workhorse when the health
check listener is enabled because the health check listener needs to
receive the SIGTERM immediately in order to set its readiness probe to
"not ready". Workhorse has a `gracefulShutdownDelay` setting that should
be used instead to drain existing connections.

This fixes the issue reported in
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/20745.

Changelog: fixed