Make Workhorse gracefulShutdownDelay default to shutdown.blackoutSeconds
With https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/4586,
we are disabling the `preStop` hook for Workhorse when the health
check listener is enabled so that Workhorse can signal immediately
that the pod is no longer ready.

However, we still need to observe the blackout time to ensure ongoing
requests complete. Previously this would require the admin to set the
`gracefulShutdownDelay` properly, but we should just observe the
blackout time if it's available.