Add openbao auth token secret file path
Prior to this change, we had no way to
provide the OpenBao authentication secret token
file path in the GitLab Helm chart.

This change fixes the issue by adding a new
configuration option in the Helm chart to specify
the file path for the OpenBao authentication
secret token.

References:
- [Rails Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/554056)
- [Helm charts issue](https://gitlab.com/gitlab-org/cloud-native/charts/openbao/-/issues/21)
- [Epic](https://gitlab.com/groups/gitlab-org/-/epics/18974)
- [ADR](https://handbook.gitlab.com/handbook/engineering/architecture/design-documents/secret_manager/decisions/011_http_audit/)
- [Rails MR](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/205780)
- [Charts MR](https://gitlab.com/gitlab-org/cloud-native/charts/openbao/-/merge_requests/36)

Changelog: added
EE : true