feat: Enable automatic DB reindexing using CronJob
Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/work_items/21668

[Automatic database reindexing] is possible in an Omnibus installation by setting the configuration
variable `gitlab_rails['database_reindexing']['enable']` to `true` in the configuration ( that is,
the `gitlab.rb` file). Similar automation is not possible in the Helm chart installation, which uses
CNG (Cloud Native GitLab) container images. There is no native method for runnig reindexing
periodically, thus, users have to resort to running this manually in the Toolbox pod or build their
own CronJob using the `toolbox` CNG container image.

The Toolbox chart already has a CronJob for taking a backup periodically. In line with that pattern,
this patch adds a CronJob that will perform [Automatic database reindexing] on a user-specified
schedule. The CronJob can be installed using these values:

```yaml
gitlab:
  toolbox:
    databaseReindex:
      cron:
        enabled: true
        schedule: "12 * * * 0,6"
```

The script used in this MR was added to CNG in
https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/2712 Before running the script, we ensure
that the dependencies of this job (Postgres and Redis) are accessible.

The underlying script will run [a Rake task], which emulates the behavior in Omnibus.

The database reindexing CronJob does not need variables related to persistence or backup. The GitLab
Geo team have [confirmed] that the reindexing does not need to run on the secondary site, as the Geo
tracking database is excluded from the [reindexing Rake task].

[Automatic database reindexing]: https://docs.gitlab.com/omnibus/settings/database/#automatic-database-reindexing
[a Rake task]: https://gitlab.com/gitlab-org/omnibus-gitlab/-/blob/b26f58eb8a5e14be5ee3ddc6c16a8ecf9a462181/files/gitlab-cookbooks/gitlab/recipes/database_reindexing_enable.rb#L27
[confirmed]: https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/4665#note_2925745040
[reindexing Rake task]: https://gitlab.com/gitlab-org/gitlab/blob/700db0f103651e068f1376d41ba52961eaed96bb/lib/tasks/gitlab/db.rake#L358-370

Changelog: added